update Design in AdminDashboard and add validation and handle exception

+++ Java_Api_Coding/CarRemarketing/src/main/java/com/gateway/traineeproject/service/ModelMasterService.java
@@ -9,11 +9,14 @@ import java.util.Optional;
 import javax.persistence.EntityExistsException;
 import javax.persistence.EntityNotFoundException;
 
+import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.gateway.traineeproject.model.Brand;
 import com.gateway.traineeproject.model.ModelMaster;
 import com.gateway.traineeproject.repository.ModelRepository;
+import com.gateway.traineeproject.repository.VehicleRepository;
 
 /**
  * @author Hiren Khatri
@@ -23,6 +26,8 @@ import com.gateway.traineeproject.repository.ModelRepository;
 public class ModelMasterService {
 	@Autowired
 	private ModelRepository modelRepository;
+	@Autowired
+	private VehicleRepository vehicleRepository;
 	
 	
 	public List<ModelMaster> findAll() {
@@ -37,6 +42,12 @@ public class ModelMasterService {
 		if(modelRepository.existsById(modelToBeSaved.getId())) {
 			throw new EntityExistsException("There is already existing entity with such ID in the database.");
 		}
+		Brand brand = new Brand();
+		brand.setId(modelToBeSaved.getBrandId());
+		if (modelRepository.existsByNameAndModelYearAndBrand(modelToBeSaved.getName(),modelToBeSaved.getModelYear(),brand)) {
+			throw new ConstraintViolationException("There is already existing entity with same name in the database.",
+					null, null);
+		}
 		return modelRepository.save(modelToBeSaved);
 	}
 	
@@ -44,12 +55,24 @@ public class ModelMasterService {
 		if(!modelRepository.existsById(modelToBeUpdated.getId())) {
 			throw new EntityNotFoundException("There is no existing entity with such ID in the database.");
 		}
+		Brand brand = new Brand();
+		brand.setId(modelToBeUpdated.getBrandId());
+		if (modelRepository.existsByNameAndModelYearAndBrand(modelToBeUpdated.getName(),modelToBeUpdated.getModelYear(),brand)
+				&& !modelRepository.existsByIdAndName(modelToBeUpdated.getId(), modelToBeUpdated.getName())) {
+			throw new EntityExistsException("There is already existing entity with same name in the database.");
+		}
 		return modelRepository.save(modelToBeUpdated);
 	}
 	
 	public void delete(Long id) {
 		if(!modelRepository.existsById(id)) {
 			throw new EntityNotFoundException("There is no existing entity with such ID in the database.");
+		}
+		ModelMaster modelMaster = new ModelMaster();
+		modelMaster.setId(id);
+		if(vehicleRepository.existsByModelMaster(modelMaster)) {
+			throw new ConstraintViolationException("Model can not be deleted as vehicles are associated with this model!",
+					null, null);
 		}
 		modelRepository.deleteById(id);
 	}
